Update environment configuration, composer dependencies, and remove obsolete SQL file
- Updated `.env.example` to reflect new environment variables.
- Modified `composer.json` and `composer.lock` to include updated dependencies.
- Removed obsolete SQL file `store-12142025.sql` to clean up the project.
- Added debug log entries for better tracking of database migration issues.
- Refactored various API controllers and requests for improved functionality and maintainability.

+++ app/Http/Requests/PurchaseOrderReceiveRequest.php
@@ -15,7 +15,7 @@ class PurchaseOrderReceiveRequest extends FormRequest
     {
         return [
             'receipts' => ['required', 'array'],
-            'receipts.*' => ['integer', 'min:0'],
+            'receipts.*' => ['numeric', 'min:0'],
             'costs' => ['sometimes', 'array'],
             'costs.*' => ['nullable', 'numeric', 'min:0'],
             'notes' => ['nullable', 'string'],
